\name{calcul_densitate}
\alias{calcul_densitate}
\title{Calculul densitatii unui sir de parametri}
\usage{
calcul_densitate(l, r, dim, parametri)
}
\description{
Primeste intervalul si dimensiunea din care se va crea variabila continua aleatoare si va calcula densitatea unei liste de parametri primiti de asemenea ca argument. In final, afiseaza graficul densitate/parametru.
}
\examples{
calcul_densitate(1,10,5,c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
