\name{calcul_medie_dispersie}
\alias{calcul_medie_dispersie}
\title{Calculeaza media si dispersia unei variabile aleatoare si functiei asociate}
\usage{
calcul_medie_dispersie(l, r, dim, ecuatie)
}
\description{
Primeste intervalul si dimensiunea din care se va crea variabila continua aleatoare si functia care se va aplica elementelor variabilei si va afisa media si dispersia acesteia.
}
\examples{
calcul_medie_dispersie(1,10,5, function(x) x * x)
}
